+++ src/classes/CodeGenerators/property/XMLPropertyGenerator.ts
@@ -1,6 +1,6 @@
 import { IPropertyGenerator } from "./interfaces/IPropertyGenerator";
 import { IPropertyGetterStrategy } from "./interfaces/IPropertyGetterStrategy";
-
+import vscode from "vscode";
 export class XMLPropertyGenerator implements IPropertyGenerator {
 	public generateProperties(strategy: IPropertyGetterStrategy) {
 		let properties: string = "";
@@ -10,9 +10,13 @@ export class XMLPropertyGenerator implements IPropertyGenerator {
 			properties += `    ${propertyTexts.name}="${propertyTexts.defaultValue}"\n`;
 		});
 
-		const parent = strategy.getParent();
-		if (parent) {
-			properties += this.generateProperties(parent);
+		const shouldAddInheritedProperties = vscode.workspace.getConfiguration("ui5.plugin").get("addInheritedPropertiesAndAggregations");
+
+		if (shouldAddInheritedProperties) {
+			const parent = strategy.getParent();
+			if (parent) {
+				properties += this.generateProperties(parent);
+			}
 		}
 
 		return properties;